# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Career Guidance System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Student {
  id: number;
  fullName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  department: string;
  semester: number;
  /** @nullable */
  percentage?: number | null;
  /** @nullable */
  attendance?: number | null;
  /** @nullable */
  aptitudeScore?: number | null;
  /** @nullable */
  communicationScore?: number | null;
  /** @nullable */
  technicalScore?: number | null;
  createdAt?: string;
}

export interface StudentInput {
  fullName: string;
  email: string;
  phone?: string;
  department: string;
  semester: number;
  percentage?: number;
  attendance?: number;
  aptitudeScore?: number;
  communicationScore?: number;
  technicalScore?: number;
}

export interface StudentUpdate {
  fullName?: string;
  email?: string;
  phone?: string;
  department?: string;
  semester?: number;
  percentage?: number;
  attendance?: number;
  aptitudeScore?: number;
  communicationScore?: number;
  technicalScore?: number;
}

export interface MlModel {
  id: number;
  modelName: string;
  modelType: string;
  /** @nullable */
  description?: string | null;
}

export interface MlModelInput {
  modelName: string;
  modelType: string;
  description?: string;
}

export interface Prediction {
  id: number;
  studentId: number;
  predictedScore: number;
  performanceLevel: string;
  student?: Student;
}

export interface PredictionInput {
  studentId: number;
}

export interface Eligibility {
  id: number;
  studentId: number;
  attendanceStatus: string;
  aptitudeStatus: string;
  percentageStatus: string;
  overallEligibility: string;
  student?: Student;
}

export interface EligibilityInput {
  studentId: number;
}

export interface CareerSuggestion {
  id: number;
  studentId: number;
  suggestedDomain: string;
  confidenceScore: number;
  student?: Student;
}

export interface CareerSuggestionInput {
  studentId: number;
}

export interface PercentageGroup {
  id: number;
  studentId: number;
  percentageRange: string;
  student?: Student;
}

export interface DashboardStats {
  totalStudents: number;
  totalPredictions: number;
  totalEligible: number;
  totalSuggestions: number;
  avgPercentage: number;
  avgAttendance: number;
  avgAptitude: number;
}

export interface DepartmentBreakdown {
  department: string;
  count: number;
}

export interface PerformanceDistribution {
  level: string;
  count: number;
}

export interface CareerDomainDistribution {
  domain: string;
  count: number;
}

export interface EligibilitySummary {
  eligible: number;
  notEligible: number;
  pending: number;
}

export type ListStudentsParams = {
department?: string;
semester?: number;
search?: string;
};