# lib/api-zod/src/generated/types/percentageGroup.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Career Guidance System API
 * OpenAPI spec version: 0.1.0
 */
import type { Student } from './student';

export interface PercentageGroup {
  id: number;
  studentId: number;
  percentageRange: string;
  student?: Student;
}