# lib/api-zod/src/generated/types/prediction.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Career Guidance System API
 * OpenAPI spec version: 0.1.0
 */
import type { Student } from './student';

export interface Prediction {
  id: number;
  studentId: number;
  predictedScore: number;
  performanceLevel: string;
  student?: Student;
}