# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Career Guidance System API
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: number;
  fullName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  department: string;
  semester: number;
  /** @nullable */
  percentage?: number | null;
  /** @nullable */
  attendance?: number | null;
  /** @nullable */
  aptitudeScore?: number | null;
  /** @nullable */
  communicationScore?: number | null;
  /** @nullable */
  technicalScore?: number | null;
  createdAt?: string;
}